made default exponential backoff more sane
2777h46m40s was the default wait.

also, most implementations of exponential backoff used for non-collision
based issues that I've seen simply use a factor of 2 (this was a factor
of N) that are seeded at some interval in [milli]seconds.

I think this was the intent with the conditional being 60 seconds.

The func for exponential backoff mentions that people should use their
own implementation. I guess this expects users to implement the worker
interface's method with their own polling algo? A sane default sounds
nice, but if the default is intended to be implement your own shouldn't
we pass a `func (time.Duration) time.Duration` as a param?